+++ packages/core/src/index.ts
@@ -308,6 +308,6 @@ export {
 } from './tokens'
 
 // ---------- 版本信息 ----------
-export const SCHEMA_UI_VERSION = '1.17.6'
+export const SCHEMA_UI_VERSION = '1.18.1'
 export const EP_FORK_BASE = '2.14.2'
 export const EP_FORK_DATE = '2026-08-27'